requirements.txt
streamlit
python-dotenv
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers
